refactor: add hooks file and use register from node:module.
